fix(progression-bank): Fixed toJson() errors and added relevant tests.

+++ lib/state/variation_group.dart
@@ -22,7 +22,7 @@ class SavedVariationGroup {
         dryVariationId = entry.progression.dryVariationId;
 
   SavedVariationGroup.fromJson(Map<String, dynamic> json)
-      : ids = json['ids'],
+      : ids = json['ids'].cast<int>(),
         dryVariationId = json['dry'];
 
   Map<String, dynamic> toJson() => {
